+++ www/codemirror/codemirror-range-main.js
@@ -2,5 +2,6 @@ const RangeHelper = {}
 
 RangeHelper.rangeDebugging = require('./codemirror-range-debugging')
 RangeHelper.rangeDetermination = require('./codemirror-range-determination')
+RangeHelper.rangeObjects = require('./codemirror-range-objects')
 
 window.RangeHelper = RangeHelper
